safari-600.7-branch/LayoutTests/platform/mac/accessibility/estimated-progress-expected.txt


This test makes sure the loading progress of a webpage can be retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.role is 'AXRole: AXWebArea'
PASS obj.isAttributeSupported('AXLoadingProgress') is true
PASS successfullyParsed is true

TEST COMPLETE
